requirements.txt
Black
fastapi[all]
fastapi_utils
fire
passlib[bcrypt]
pylint
python-decouple
SQLAlchemy



